don't use `get_all_contents` every process for family heirlooms (#92576)
## About The Pull Request

the family heirloom quirk runs `get_all_contents` on the quirk holder
every process. which is kinda dumb and wasteful when we are looking to
see if the quirk holder has an item _that we already have a reference
to_... `/atom/proc/contains` exists and works just fine for this.

## Why It's Good For The Game

`get_all_contents` is very wasteful and unneeded for this use-case

## Changelog

no user-facing changes