Fixes tank holder description (#92594)
## About The Pull Request
Fixes "It is The bolts[...]" in tank holder desc, changes tanks from
"one" to "a"/"an" (i.e. "an oxygen tank" instead of "one oxygen tank")
and adds span_notice to the other parts of the description to fit with
the screws message.
## Why It's Good For The Game
Fixes a grammar mistake in the description.
The "one [...]" doesn't make sense for something that only holds a
single tank at a time.
Using span_notice for the description makes it more uniform and
consistent.
## Changelog
:cl:
spellcheck: Fixes "It is The bolts on the bottom are..." in tank holder
descriptions, and improves grammar for inserted objects.
/:cl:

---------

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>