Gives blood to heretic victims upon arriving in mansus (#92744)
## About The Pull Request

Restores victims of a heretic sacrifice's blood to blood_volume_okay if
they need it upon arriving in the mansus.
Does not conflict with #92119.

## Why It's Good For The Game

closes #88557
As was said in the issue, dying because you were too asleep to do
anything about it is lame. The determination buff already gives you
quite strong blood regen, but it's the initial value that counts.