Makes the slot machine spin faster (#92876)
## About The Pull Request

Does what it says, taking the slot machine from 6 seconds to 2 seconds
to spin.

## Why It's Good For The Game

Was playing a round of tram and was annoyed at how slow it was to
gamble. It also makes sense the slot machine goes quick so you waste
your money faster.

:cl: Webcomicartist
qol: The slot machine spins faster, allowing for more time-efficient
gambling.
/:cl: