Buffs the amount EMPs drain cells to be worthwhile (#92902)
Simply put, EMPs now drain cells at hopefully decent rates. Starting at
basic high capacity, a heavy emp (aka a direct ion rifle hit) drains 1/3
of the battery, with each successive tier of cell taking an extra heavy
emp to drain it. At the end, the data looks like this:

- Basic Cell: Instantly drained. Unchanged.
- High Capacity: 3 hits
- Super Capacity: 4 Hits
- Hyper Capacity: 6 Hits
- Bluespace: 8 hits.

Also potatoes take halfed emp damage, because theres not really much
silicon to emp, given its a potato.
## Why It's Good For The Game

As of now, this is the stats to drain a cell with EMPs. (Do note some
stuff just doesnt drain at all and this is for those that do just run
the proc to the cell itself)

- Basic Cell: Instantly drained.
- High Capacity: 10 hits
- Super Capacity: 20 Hits
- Hyper Capacity: 30 Hits
- Bluespace: 40 hits.

The EMP gun on the station can shoot 10 bolts before needing a recharge.
It is *physically impossible* to use the ion to depower something with
an upgraded cell without needing to reload, and even landing 10 direct
hits with the ion to bring down a stock cell is frankly a big ask.
Additionally, using emps as crew or sec runs the risk of you getting
caught in the splash damage and now your modsuit/pda/whatever doesnt
work. Hopefully, this will make the EMP gun actually useful for its
intended purpose of countering silicons/mechs.

## Changelog
:cl: WebcomicArtist
balance: EMPs now drain cells better, starting at 3 emps to drain a
basic cell, up to 8 for bluespace.
/:cl:

---------

Co-authored-by: Aki Ito <11748095+ExcessiveUseOfCobblestone@users.noreply.github.com>