Fix keycombo Modal not working after clear (#93146)
## About The Pull Request
Keycombo has a button to clear the assigned key. Doing so says that one
can press a key to record it. The way autofocus works though is that
this only ever works once on the first window opening. Not if one wants
to reassign a hotkey.
## Why It's Good For The Game
## Changelog
:cl:
fix: keycombo modal only allowing an input on the first opening
/:cl: