Things that are edible now count as golem food if their materials allow it. (#92582)
## About The Pull Request
This PR allows golems or anyone with the genetic mutation to benefit
from the effects of golem food if what they're eating is made of a
material tied to ore or sheets with golem food effects.

In layman terms, this means you can deepfry a chair and get golem buffs
that way. The duration also scales with the amount of material consumed,
which also depends on the bite size of the food. Smaller doses translate
to a shorter duration of the effect, while the viceversa is also true
(however with diminishing returns).

Also, you won't take immediate damage from biting, food with items made
of metals and glass stuffed into it if you've the rock eater trait (come
on dude, you eat literal raw diamonds). This has nothing to do with
golem food, but it's one less peeve off the list as I slowly and
occasionally work on material datums and their presence in the game.

## Why It's Good For The Game
Adding some depth to materials and food, albeit not much.