Non-admin spawned atlas gravitronic spinal implants now use their intentional(?) sprite (#93428)
## About The Pull Request

so apparently there are two sprites for herculean gravitronics, one with
the core in and one without, but the code for adding a core to the
gravitronic doesn't actually change the sprite so it's unused unless an
admin spawns in an atlas with a core already in it which in that case
it'll have the sprite.

This single line of code makes the sprite change when a new core is put
in.

yes I tested it no it's not a web edit

## Why It's Good For The Game

bugfix

## Changelog

:cl:
fix: Atlas gravitronic spinal implants now use their intended sprite
/:cl:

Co-authored-by: Fghj240 <fakeemail@notrealemail.com>